+++ src/Symfony/Component/Uid/UuidV1.php
@@ -41,6 +41,18 @@ class UuidV1 extends Uuid implements TimeBasedUidInterface
         return uuid_mac($this->uid);
     }
 
+    public function toV6(): UuidV6
+    {
+        $uuid = $this->uid;
+
+        return new UuidV6(substr($uuid, 15, 3).substr($uuid, 9, 4).$uuid[0].'-'.substr($uuid, 1, 4).'-6'.substr($uuid, 5, 3).substr($uuid, 18, 6).substr($uuid, 24));
+    }
+
+    public function toV7(): UuidV7
+    {
+        return $this->toV6()->toV7();
+    }
+
     public static function generate(\DateTimeInterface $time = null, Uuid $node = null): string
     {
         $uuid = !$time || !$node ? uuid_create(static::TYPE) : parent::NIL;