+++ src/Symfony/Component/Uid/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 7.1
 ---
 
- * Add `UuidV6::fromV1()` and `UuidV7::fromV1()`
+ * Add `UuidV1::toV6()`, `UuidV1::toV7()` and `UuidV6::toV7()`
 
 6.2
 ---